Separate "Simple Input" for the new input changes
