+++ src/components/input/input-simple/InputSimple.module.scss.d.ts
@@ -1,2 +1,1 @@
 export declare const aminoInputWrapper: string;
-export declare const fields: string;
