+++ src/components/input/input-simple/input-type/_NumberInput.module.scss.d.ts
