+++ src/components/input/input-simple/input-type/_PasswordInput.module.scss
@@ -1,0 +1,22 @@
+@use 'theme';
+
+.styledWrapper {
+  position: relative;
+  width: 100%;
+}
+
+.styledButtonAction.styledButtonAction {
+  padding: 6px;
+  border-radius: 50%;
+  transition: theme.$amino-transition;
+
+  &:hover {
+    background: theme.$amino-hover-color;
+  }
+}
+
+.aminoInput {
+  input {
+    padding-right: theme.$amino-space-40;
+  }
+}
