+++ src/components/input/input-simple/input-type/_PasswordInput.module.scss.d.ts
@@ -1,4 +1,3 @@
 export declare const aminoInput: string;
-export declare const styledActionWrapper: string;
 export declare const styledButtonAction: string;
 export declare const styledWrapper: string;
