+++ src/components/input/input-type/_TimeInput.module.scss.d.ts
@@ -1,4 +1,4 @@
-export declare const input: string;
+export declare const aminoInput: string;
 export declare const styledActionWrapper: string;
 export declare const styledButtonAction: string;
 export declare const styledWrapper: string;
